% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelism.R
\name{parallelism}
\alias{parallelism}
\title{Interaction test for bivariate smoothing spline ANOVA}
\usage{
parallelism(y = y, z = z, g = g)
}
\arguments{
\item{y}{the input signal.}

\item{z}{the spatial or tempral locations (scaled to [0,1]).}

\item{g}{the categorical variable representing the group information.}
}
\value{
score the value of the test statisitcs

pvalue P-value of the test
}
\description{
Interaction test for bivariate smoothing spline ANOVA
}
\examples{
\donttest{
library(gss)
library(Parallelism)
n = 200
##generate locataions of the spatial or temperal points
z= seq(0,1,length.out=n)
z = c(z,z)
##the group indictor for each mearsuement
g = c(rep(0,n),rep(1,n))

##generate the measured signal for each group at different locations
fun1 = function(x,delta) 2.5*sin(3*pi*x[,2])*(1-x[,2])*ifelse(x#' [,1]==0,1,0) + ( (2.5+delta)*sin(3*pi*x[,2])*(1-x[,2]) )*#' ifelse(x[,1]==0,0,1)
y = fun1(df, 0.5) + rnorm(2*n)

##begin the test
res = parallelism(y,z,g)
}
}
